+++ st2common/st2common/models/api/keyvalue.py
@@ -20,12 +20,14 @@ from keyczar.keys import AesKey
 from oslo_config import cfg
 import six
 
-from st2common.exceptions.keyvalue import CryptoKeyNotSetupException
+from st2common.constants.keyvalue import SYSTEM_SCOPE, USER_SCOPE, ALLOWED_SCOPES
+from st2common.exceptions.keyvalue import CryptoKeyNotSetupException, InvalidScopeException
 from st2common.log import logging
 from st2common.util import isotime
 from st2common.util import date as date_utils
 from st2common.util.crypto import symmetric_encrypt, symmetric_decrypt
 from st2common.models.api.base import BaseAPI
+from st2common.models.system.keyvalue import UserKeyReference
 from st2common.models.db.keyvalue import KeyValuePairDB
 
 LOG = logging.getLogger(__name__)
@@ -62,6 +64,11 @@ class KeyValuePairAPI(BaseAPI):
                 'type': 'boolean',
                 'required': False,
                 'default': False
+            },
+            'scope': {
+                'type': 'string',
+                'required': False,
+                'default': SYSTEM_SCOPE
             },
             'expire_timestamp': {
                 'type': 'string',
@@ -111,19 +118,26 @@ class KeyValuePairAPI(BaseAPI):
 
         doc = cls._from_model(model, mask_secrets=mask_secrets)
 
-        if 'id' in doc:
-            del doc['id']
-
-        if model.expire_timestamp:
+        if getattr(model, 'expire_timestamp', None) and model.expire_timestamp:
             doc['expire_timestamp'] = isotime.format(model.expire_timestamp, offset=False)
 
         encrypted = False
-        if model.secret:
+        secret = getattr(model, 'secret', False)
+        if secret:
             encrypted = True
 
-        if not mask_secrets and model.secret:
+        if not mask_secrets and secret:
             doc['value'] = symmetric_decrypt(KeyValuePairAPI.crypto_key, model.value)
             encrypted = False
+
+        scope = getattr(model, 'scope', SYSTEM_SCOPE)
+        if scope:
+            doc['scope'] = scope
+
+        key = doc.get('name', None)
+        if scope == USER_SCOPE and key:
+            doc['user'] = UserKeyReference.get_user(key)
+            doc['name'] = UserKeyReference.get_name(key)
 
         doc['encrypted'] = encrypted
         attrs = {attr: value for attr, value in six.iteritems(doc) if value is not None}
@@ -134,6 +148,7 @@ class KeyValuePairAPI(BaseAPI):
         if not KeyValuePairAPI.crypto_setup:
             KeyValuePairAPI._setup_crypto()
 
+        kvp_id = getattr(kvp, 'id', None)
         name = getattr(kvp, 'name', None)
         description = getattr(kvp, 'description', None)
         value = kvp.value
@@ -145,16 +160,24 @@ class KeyValuePairAPI(BaseAPI):
         else:
             expire_timestamp = None
 
-        if getattr(kvp, 'secret', False):
+        secret = getattr(kvp, 'secret', False)
+
+        if secret:
             if not KeyValuePairAPI.crypto_key:
                 msg = ('Crypto key not found in %s. Unable to encrypt value for key %s.' %
                        (KeyValuePairAPI.crypto_key_path, name))
                 raise CryptoKeyNotSetupException(msg)
             value = symmetric_encrypt(KeyValuePairAPI.crypto_key, value)
-            secret = True
 
-        model = cls.model(name=name, description=description, value=value,
-                          secret=secret,
+        scope = getattr(kvp, 'scope', SYSTEM_SCOPE)
+
+        if scope not in ALLOWED_SCOPES:
+            raise InvalidScopeException('Invalid scope "%s"! Allowed scopes are %s.' % (
+                scope, ALLOWED_SCOPES)
+            )
+
+        model = cls.model(id=kvp_id, name=name, description=description, value=value,
+                          secret=secret, scope=scope,
                           expire_timestamp=expire_timestamp)
 
         return model