+++ st2common/st2common/services/keyvalues.py
@@ -13,16 +13,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from st2common.constants.keyvalue import SYSTEM_SCOPE, USER_SCOPE, ALLOWED_SCOPES
+from st2common.exceptions.keyvalue import InvalidScopeException, InvalidUserException
+from st2common.models.system.keyvalue import UserKeyReference
 from st2common.persistence.keyvalue import KeyValuePair
 
 
 class KeyValueLookup(object):
 
-    def __init__(self, key_prefix='', cache=None):
+    def __init__(self, key_prefix='', cache=None, scope=SYSTEM_SCOPE):
         self._key_prefix = key_prefix
         if cache is None:
             cache = {}
         self._value_cache = cache
+        self._scope = scope
 
     def __str__(self):
         return self._value_cache[self._key_prefix]
@@ -42,15 +46,72 @@ class KeyValueLookup(object):
         # the lookup is for 'key_base.key_value' it is likely that the calling code, e.g. Jinja,
         # will expect to do a dictionary style lookup for key_base and key_value as subsequent
         # calls. Saving the value in cache avoids extra DB calls.
-        return KeyValueLookup(key, self._value_cache)
+        return KeyValueLookup(key_prefix=key, cache=self._value_cache, scope=self._scope)
 
     def _get_kv(self, key):
-        kvp = None
-        try:
-            kvp = KeyValuePair.get_by_name(key)
-        except ValueError:
-            # ValueErrors are expected in case of partial lookups
-            pass
-        # A good default value for un-matched value is empty string since that will be used
-        # for rendering templates.
+        scope = self._scope
+        kvp = KeyValuePair.get_by_scope_and_name(scope=scope, name=key)
         return kvp.value if kvp else ''
+
+
+class UserKeyValueLookup(object):
+
+    def __init__(self, user, key_prefix='', cache=None, scope=USER_SCOPE):
+        self._key_prefix = key_prefix
+        if cache is None:
+            cache = {}
+        self._value_cache = cache
+        self._user = user
+        self._scope = scope
+
+    def __str__(self):
+        return self._value_cache[self._key_prefix]
+
+    def __getitem__(self, key):
+        return self._get(key)
+
+    def __getattr__(self, name):
+        return self._get(name)
+
+    def _get(self, name):
+        # get the value for this key and save in value_cache
+        if not self._key_prefix:
+            key = UserKeyReference(name=name, user=self._user).ref
+        else:
+            key = '%s.%s' % (self._key_prefix, name)
+        value = self._get_kv(key)
+        self._value_cache[key] = value
+        # return a KeyValueLookup as response since the lookup may not be complete e.g. if
+        # the lookup is for 'key_base.key_value' it is likely that the calling code, e.g. Jinja,
+        # will expect to do a dictionary style lookup for key_base and key_value as subsequent
+        # calls. Saving the value in cache avoids extra DB calls.
+        return UserKeyValueLookup(user=self._user, key_prefix=key, cache=self._value_cache)
+
+    def _get_kv(self, key):
+        scope = self._scope
+        kvp = KeyValuePair.get_by_scope_and_name(scope=scope, name=key)
+        return kvp.value if kvp else ''
+
+
+def get_key_reference(scope, name, user=None):
+    """
+    Given a key name and user this method returns a new name (string ref)
+    to address the key value pair in the context of that user.
+
+    :param user: User to whom key belongs.
+    :type name: ``str``
+
+    :param name: Original name of the key.
+    :type name: ``str``
+
+    :rtype: ``str``
+    """
+    if scope == SYSTEM_SCOPE:
+        return name
+    elif scope == USER_SCOPE:
+        if not user:
+            raise InvalidUserException('A valid user must be specified for user key ref.')
+        return UserKeyReference(name=name, user=user).ref
+    else:
+        raise InvalidScopeException('Scope "%s" is not valid. Allowed scopes are %s.' %
+                                    (scope, ALLOWED_SCOPES))